WIP added basic explore map with clusters
