+++ app/explore/page.tsx
@@ -1,0 +1,6 @@
+import Link from 'next/link';
+import ExploreMap from '@/components/ExploreMap'; 
+
+const Explore = () => <ExploreMap />
+
+export default Explore;
