+++ lib/constants.ts
@@ -1,0 +1,4 @@
+export const MAP_STYLES = {
+    dark: 'mapbox://styles/mapbox/dark-v11',
+    standard: 'mapbox://styles/mapbox/standard',
+  };
