+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@types/react-map-gl": "^6.1.6",
     "mapbox-gl": "^3.4.0",
     "next": "14.2.3",
     "react": "^18",
